Renamed list to datastructures and moved it
